Split code into seperate file
