+++ Toolkit/Toolkit.Planners/DepthFirstSearch.cs
@@ -78,23 +78,5 @@ namespace PDDLSharp.Toolkit.Planners
 
             return new ActionPlan(bestGoal[0].Steps, bestGoal[0].Steps.Count);
         }
-
-        internal class StateMove
-        {
-            public IState State { get; private set; }
-            public List<GroundedAction> Steps { get; private set; }
-
-            public StateMove(IState state, List<GroundedAction> steps)
-            {
-                State = state;
-                Steps = steps;
-            }
-
-            public StateMove(IState state)
-            {
-                State = state;
-                Steps = new List<GroundedAction>();
-            }
-        }
     }
 }
